--- a/mysql/makefile
+++ b/mysql/makefile
@@ -1,19 +1,18 @@
-build:
+build: clean
 	docker build -t mysql-volumeless:8.0 .
-
-run:
-	docker run --name mysql-volumeless \
+	docker run -d \
+	       	--name mysql-volumeless \
 		-e MYSQL_USER=moov \
 		-e MYSQL_PASSWORD=secret \
 		-e MYSQL_ROOT_PASSWORD=secret \
+		-e MYSQL_DATABASE=test \
+		-p 3306:3306 \
 	       	mysql-volumeless:8.0
-
-stop:
+	./test-mysql-is-ready.sh
 	docker stop mysql-volumeless
-
-image:
 	docker commit mysql-volumeless mysql-volumeless:8.0
-	docker tag mysql-volumeless:8.0 vaulty/mysql-volumeless:8.0
-
+	docker tag mysql-volumeless:8.0 moov/mysql-volumeless:8.0
 push:
-	docker push vaulty/mysql-volumeless:8.0
+	docker push moov/mysql-volumeless:8.0
+clean:
+	docker rm mysql-volumeless