@adamdecaf @vxio 

With this change:
1. We run single MySQL container that all tests use
2. Each test that calls `db := CreateTestMySQLDB` gets own unique database. On `db.Close()`database is dropped.
3. We don't stop / remove MySQL container as we don't know when all out tests are done

- [x] I created custom volume-less MySQL container that starts in 3s instead of 12s. I host it on own DockerHub public repo. It would be better to move under Moov's DockerHub account. @adamdecaf 

Using this changes reduces tests time for
* Customer's from 2m:30s to 28s (on my machine)
* Paygate's  from to 1m:10s to 30s (on my machine)