@vxio 2020-10-29
@alovak I recommend working closely with @adamdecaf on this since he has still has #101 open.

Where do we teardown/cleanup the database for each test?