@alovak 2020-10-29
@vxio I'm working on it because we agreed with Adam that I'll take his PR and re-apply it on top of master. There are some parts missed (like creating random db, etc.) that's why you see my PR and some parts were changed.

> Where do we teardown/cleanup the database for each test?

This is I want to discuss with you both @vxio @adamdecaf 

There are two aspects:
1. remove temporary DB that was created for specific test
2. shut down MySQL server instance

1st is pretty simple. When you call db.Close() (db is TestMySQLDB) we can `drop database xxx`
2nd is harder as it should wait until `go test` finishes or ... as compromise:
- we can shut it down manually in TestMain()
  - benefit is that all tests in directory (and all parallel tests) will use shared MySQL server
  - drawback is that we have to manually call database.SharedInstance.Shutdown()
- we can run parallel tests (t.Parallel()) then we may hope that most of them will use same shared MySQL server
 - benefits we get faster tests
 - drawbacks - we may need to update some tests (make them ready for concurrent/parallel run)
 - parallel tests still may make 0 waiters where instance will be shut down and launched again