@adamdecaf 2020-10-29
Are we making this too complicated on ourselves when we should be using an ORM? I'm not sure we can solve this in Go without BeforeAll/AfterAll, which would be really nice but doesn't exist.