@vxio 2020-10-29
> Are we making this too complicated on ourselves when we should be using an ORM? I'm not sure we can solve this in Go without BeforeAll/AfterAll, which would be really nice but doesn't exist.

An ORM would be nice if we still want to support sqlite (and other dbs eventually) but it wouldn't fix our issue with slow tests when using `mysql`