@alovak 2020-10-31
@vxio @adamdecaf in the latest change we don't shutdown MySQL container at all.

As there is no exit trigger when we can shutdown MySQL we may just keep it running. It's not ideal solution, but so far I think it's good:
* as developer I run tests multiple times so I still need it
* one container is not a lot :D (before I could have many running mysql containers because sometimes they were not shutdown for some reasons)
* on CI container will be shutdown by external system
* ...

any better ideas?