@adamdecaf 2020-11-01
That's not idea, but I understand we don't have a solid way to detect all tests have finished. I was hoping we could leverage `sync.WaitGroup`, but that doesn't seem to work with sequential tests.