@vxio 2020-11-04
Here are some suggestions for remove the global variables and simplifying the configuration:

https://github.com/moov-io/base/pull/121