@alovak 2020-11-06
@adamdecaf @vxio let's merge it until it's fresh :)

We can do more improvements later if needed.