@alovak 2020-11-11
@adamdecaf now we have moov/mysql-volumeless

I created workflow to build and publish image. It's triggered on created tag v*