@alovak 2020-11-11
@vxio tried to find "best practices" on using/passing testing.T in helper and was not able to find anything that would make me want to pass it to `findOrLaunchMySQLContainer` and `createTemporaryDatabase` functions :D 

Also, I found this [here](https://dave.cheney.net/2019/12/08/dynamically-scoped-variables-in-go):

> This is a bad idea, no argument there. This is not a pattern you should ever use in production code. But, this isn’t production code, it’s a test, and perhaps there are different rules that apply to test code. After all, we use mocks, and stubs, and monkey patching, and type assertions, and reflection, and helper functions, and build flags, **and global variables**, all so we can test our code effectively. None of those, uh, hacks will ever show up in the production code path, so is it really the end of the world?

> If you’ve read this far perhaps you’ll agree with me that as unconventional as this approach is, not having to pass a *testing.T into every function that could possibly need to assert something transitively, makes for clearer test code.

> So maybe, in this case, the ends do justify the means.

What I want to say here is that if there is no clear standard/guideline/benefit from doing something (not doing something), let's just make it as is an move forward. Maybe next time, if we have an issue we will know more to make a better decision.