@adamdecaf 2020-11-11
It seems his solution is to [inspect the runtime stack](https://github.com/pkg/expect/blob/master/t.go#L18) instead of using stdlib solutions (IMO `t.Helper()` exists because you're supposed to pass around `*testing.T`) which could break between Go releases. His solution is absolutely not suitable for the average Go programmer to use.

We can postpone this, but IMO the Go team officially supports passing `*testing.T` and encourages it.